% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_181_reporter.R
\name{diann_181_reporter}
\alias{diann_181_reporter}
\title{DIA-NN version 1.8.1 Report Generator}
\usage{
diann_181_reporter(
  report_in = "report.tsv",
  report_out = "diann_report.tsv",
  qvalue_filter = 0.01,
  pep_filter = 0.2,
  maxlfq_group = "Protein.Group",
  maxlfq_id = "Precursor.Id",
  maxlfq_quantity = "Precursor.Normalised"
)
}
\arguments{
\item{report_in}{Name of the report.tsv from DIA-NN}

\item{report_out}{Name of the output report}

\item{qvalue_filter}{Value to filter on for the Q-Values}

\item{pep_filter}{Value to filter on for the PEP}

\item{maxlfq_group}{Column name for the protein group in MaxLFQ calculation}

\item{maxlfq_id}{Column name for the Precursor ID to use in MaxLFQ calculation}

\item{maxlfq_quantity}{Column name to use for the quantity in MaxLFQ calculation}
}
\value{
A tsv file with filtered protein abundances and number of peptides
}
\description{
Creates a .tsv report from the 'report.tsv' from DIA-NN version 1.8.1. Has
more customization options than the base reporter.
}
