% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_202_reporter.R
\name{diann_202_reporter}
\alias{diann_202_reporter}
\title{DIA-NN version 2.0 Report Generator}
\usage{
diann_202_reporter(
  report_in = "report.parquet",
  pg_matrix_in = "report.pg_matrix.tsv",
  report_out = "diann_report.tsv",
  contam_out = "diann_contaminants.tsv",
  filtered_out = "diann_poor_quality.tsv",
  qc_out = "diann_qc.tsv",
  qvalue_filter = 0.01,
  quality_filter = 0,
  pep_filter = 0.2,
  contaminants = "KRT|cRAP",
  protein_id_columns = c("Protein.Group", "Genes", "Protein.Names",
    "First.Protein.Description")
)
}
\arguments{
\item{report_in}{Name of the report.parquet from DIA-NN}

\item{pg_matrix_in}{Name of the report.pg_matrix.tsv from DIA-NN}

\item{report_out}{Name of the main output report}

\item{contam_out}{Name of the contaminant output report}

\item{filtered_out}{Name of the filtered ID output report}

\item{qc_out}{Name of the QC output report}

\item{qvalue_filter}{Value to filter on for the Q-Values}

\item{quality_filter}{Value to filter on for Quantity and PG MaxLFQ Qualities}

\item{pep_filter}{Value to filter on for the PEP}

\item{contaminants}{String to filter on for contaminants}

\item{protein_id_columns}{The name of the protein identifiers you want to
keep in your output data}
}
\value{
Four tsv files with filtered protein abundances and number of
 peptides, contaminants, filtered precursor IDs, and QC metrics
}
\description{
Creates a .tsv report from the 'report.tsv' from DIA-NN version 2.0 or
 greater. Has more customization options than the base reporter. Also creates
 more files.
}
