% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_reporter.R
\name{diann_reporter}
\alias{diann_reporter}
\title{DIA-NN Report Generator}
\usage{
diann_reporter(report_in, report_out)
}
\arguments{
\item{report_in}{Name of the report.tsv from DIA-NN}

\item{report_out}{Name of the output report}
}
\value{
A tsv file with filtered protein abundances and number of peptides
}
\description{
Creates a .tsv report from the 'report.tsv' from DIA-NN
}
