% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_report.R
\name{diann_reporter}
\alias{diann_reporter}
\alias{peptide_count}
\title{URMC MSRL DIA-NN Report Generator}
\usage{
peptide_count(df)

diann_reporter(report_in, report_out)
}
\arguments{
\item{df}{a data frame from an upstream step}

\item{report_in}{Name of report.tsv file generated from DIA-NN. Needs to be in quotes.}

\item{report_out}{Name of the output report. Can be whatever you want. Needs to be in quotes.}
}
\description{
Using the 'report.tsv' generating from a DIA-NN search, creates a new file that includes MaxLFQ and number of peptides in wide format
Requires the DiaNN R package to be installed from github (https://github.com/vdemichev/diann-rpackage) to run properly.
To install follow these steps:
install.packages("devtools")
library(devtools)
isntall_github("https://github.com/vdemichev/diann-rpackage")
}
